fastjson: Move to own pkg
